refactor(cosmic-swingset): move some helpers
