+++ packages/cosmic-swingset/src/helpers/bufferedStorage.js
