+++ packages/cosmic-swingset/src/helpers/is-entrypoint.js
