+++ packages/cosmic-swingset/src/helpers/json-stable-stringify.js
