+++ packages/cosmic-swingset/src/launch-chain.js
@@ -35,7 +35,7 @@ import {
   BeansPerXsnapComputron,
 } from './sim-params.js';
 import { parseParams } from './params.js';
-import { makeQueue } from './make-queue.js';
+import { makeQueue } from './helpers/make-queue.js';
 
 const console = anylogger('launch-chain');
 const blockManagerConsole = anylogger('block-manager');
